tests: consolidate HolidayChecker and TransactionCostManager tests, label every merged section
Five files fold into two, each named for the component it exercises rather than
the incident that produced it, finishing the pass started in 0b62ed57.

  core/test_holiday_calendar_coverage.cpp  ┐
  core/test_holiday_lookback.cpp           ├─> core/test_holiday_checker.cpp
  core/test_holiday_path_resolution.cpp    ┘

  transaction_cost/test_equity_borrow_fees.cpp            ┐   transaction_cost/
  transaction_cost/test_unknown_equity_costs_dispatch.cpp ┴─> test_transaction_cost_manager.cpp

Each former file survives as a labelled block: a banner names the SUBJECT the
block covers and why those cases exist, then that section's helpers in their own
anonymous namespace, then its tests. Concatenating without that would have made
the files cheaper to open and harder to read -- the point of grouping by
component is lost if a reader cannot tell which cases speak to which concern.

The same SUBJECT lines are added to the corp-action files merged earlier
(test_applier.cpp, test_audit_log.cpp). Those already carried provenance and a
note about namespace isolation, but nothing said what each block was for; the
dedup section in particular is the highest-stakes block in the component and now
says so. No test bodies were restructured there, only banners.

Also removes HolidayChecker::coverage_years(), which had no production caller and
existed only for three assertions in its own test. Those now assert the same
property through covers_year() and coverage_description(): the description pins
the min and max, and probing either side of the range plus both members leaves no
room for a different set, so "exactly {2024, 2025}" is still proven -- including
that an empty year counts as covered, since "no closures that year" is an answer
and "never loaded" is not.

No test lost, renamed or disabled: every TEST/TEST_F suite+name pair is identical
before and after (1385 declarations, diffed pairwise, zero on either side only).
Checked for duplicate suite+name collisions before merging, since gtest identity
is the string and a C++ namespace does not disambiguate it -- there were none.
Suite 1431/1431.
